backend changes to deployment

+++ backend/domain/src/main/java/com/xplug/tech/crop/data/Spacing.java
@@ -15,9 +15,9 @@ import javax.persistence.Enumerated;
 @Embeddable
 public class Spacing {
 
-    private Integer minimumSpacing;
+    private Double minimumSpacing;
 
-    private Integer maximumSpacing;
+    private Double maximumSpacing;
 
     @Enumerated(EnumType.STRING)
     private MeasurementUnit unit;
